Only update deck if there is a set
